feat/#313 : error flow 추가
Error 컴포넌트에 홈으로 돌아가기 버튼 추가
초대된 패킹리스트가 삭제 된 경우 error flow 추가
에러코드별 로직 추가
