+++ pages/edit-profile.tsx
@@ -1,4 +1,3 @@
-import HeadMeta from '../components/HeadMeta';
 import EditProfileLanding from '../components/profile/EditProfileLanding';
 import { AsyncBoundary } from '../utils/AsyncBoundary';
